requirements.txt
pandas==2.3.1
google-cloud==0.34.0
sqlalchemy==2.0.41
psycopg2==2.9.10
chardet==5.2.0
python-dotenv==1.1.1
google-adk
